requirements.txt
fastapi
uvicorn
pytest
pytest-cov
pymongo
python-dotenv
pika